backend/requirements.txt
fastapi
uvicorn
pydantic
qdrant-client
asyncpg
google-generativeai
pypdf
python-dotenv
tiktoken
python-docx
